requirements.txt
flask>=3.0.0
pypdf>=4.0.0
Pillow>=10.0.0
img2pdf>=0.4.4
gunicorn>=21.2.0
